refactor: move unleash env logic to useUnleashEnvironment hook
