+++ src/app/utils/experiments/UnleashProvider.tsx
@@ -1,7 +1,7 @@
-import { useDevToggle, useIsStaging } from "app/store/GlobalStore"
 import { createContext, useCallback, useEffect, useState } from "react"
 import useAppState from "../useAppState"
 import { forceFetchToggles } from "./helpers"
+import { useUnleashEnvironment } from "./hooks"
 import { getUnleashClient } from "./unleashClient"
 
 interface UnleashContext {
@@ -12,14 +12,7 @@ export const UnleashContext = createContext<UnleashContext>({ lastUpdate: null }
 
 export function UnleashProvider({ children }: { children?: React.ReactNode }) {
   const [lastUpdate, setLastUpdate] = useState<UnleashContext["lastUpdate"]>(null)
-  const isStaging = useIsStaging()
-  const unleashEnv = __DEV__
-    ? useDevToggle("DTUseProductionUnleash")
-      ? "production"
-      : "staging"
-    : isStaging
-    ? "staging"
-    : "production"
+  const { unleashEnv } = useUnleashEnvironment()
 
   useEffect(() => {
     const client = getUnleashClient(unleashEnv)
